requirements.txt
asgiref==3.3.4
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
Faker==8.8.1
marshmallow==3.12.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
unrar==0.4
webargs==8.0.0
